Changes to token gen function
