+++ api/v1/services/TokenService.js
@@ -2,6 +2,7 @@ var Token = require('../models/Token');
 var config = require('../../config');
 var logger = require('../../logging');
 var errors = require('../errors');
+var utils = require('../utils');
 
 /**
  * Finds a token given the Token value
@@ -44,3 +45,28 @@ module.exports.findTokenByValue = function(value, scope) {
 			return Promise.resolve(result);
 		});
 };
+
+/**
+ * Generates a token and deletes all existing tokens
+ * with the same scope.
+ * @param {User} user The user object to create a reset token for.
+ * @param {String} scope The scope to create the token for.
+ * @return {Promise<Bool>} Returns a Promise that resolves to
+ *                         true on a successful token creation.
+ */
+module.exports.generateToken = function(user, scope){
+	var tokenVal = utils.crypto.generateResetToken();
+	var user_id = user.get('id');
+
+	return Token
+		.where({user_id: user_id, type: scope}).fetchAll()
+		.then(function(tokens) {
+			return tokens.invokeThen('destroy')
+				.then(function() {
+					var token = Token.forge({type: scope, value: tokenVal, 'user_id': user_id});
+					return token.save().then(function () { return tokenVal; });
+				});
+		});
+};
+
+
